+++ CHANGELOG.md
@@ -4,9 +4,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+- command line argument for prefix in test-need analysis
 - some more tests added
 ### Changed
-- test-need analysis not outputs csv (instead of console output)
+- test-need analysis now outputs csv (instead of console output)
 - update to clojure 1.8.0
 
 ##[0.6.2] - 2016-03-15